feat(events): add admin event creation form and toasts
Add AdminCreateEvent component to provide a full event creation
UI for administrators. The form includes title, description, image
URL and start/end date pickers (using CustomDatePicker). Form state
is managed with useState and submissions call createEvent from the
eventService. Success and error feedback use react-hot-toast and a
Toaster.

Also add new runtime deps to package-lock: react-datepicker and
react-hot-toast, plus related lockfile entries required by those
packages.

This enables admins to create events from the frontend with visual
feedback and date selection.

+++ backend/Models/Attendee.cs
@@ -7,6 +7,6 @@ public class Attendee
     public string? Avatar { get; set; }
     public long? EmployeeId { get; set; }
     public Employee? Employee { get; set; }
-    public int EventId { get; set; }
+    public long EventId { get; set; }
     public Event? Event { get; set; }
 }
